refactoring util's modules
